[packaging] Drop Mozilla Location Service API key
We are not Debian, even though we still report as it. We can't
use the Debian-allocated MLS API key.

Unfortunately Mozilla didn't reply about a Droidian specific
key request back last year, so we don't have a replacement one.

Mozilla Location Service is being sunsetted anyway, and even
already-issued private keys will be dropped in a couple of
weeks (June 12th, 2024), so let's not even bother with all of this,
and remove the key straight away.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>